refactor(layout): separate authenticated & non-authenticated paths
this enables conditional rendering of the navbar without using client-side hooks, reducing desync between auth state and visual indicatoins
