+++ src/app/(authenticated)/error/[code]/page.tsx
