+++ src/app/(authenticated)/guide_articles/create/page.tsx
