+++ src/app/(authenticated)/guide_articles/edit/[id]/page.tsx
