+++ src/app/(authenticated)/guide_articles/layout.tsx
