+++ src/app/(authenticated)/guide_articles/page.tsx
