+++ src/app/(authenticated)/layout.tsx
@@ -1,0 +1,17 @@
+import type { ReactNode } from "react";
+
+import { ContentWrapper } from "@/components/content-wrapper";
+import { Navbar } from "@/components/navbar";
+
+export default function AuthenticatedLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return (
+    <>
+      <Navbar />
+      <ContentWrapper>{children}</ContentWrapper>
+    </>
+  );
+}
