+++ src/app/(authenticated)/student_organizations/edit/[id]/page.tsx
