+++ src/app/(authenticated)/student_organizations/layout.tsx
