+++ src/app/(authenticated)/student_organizations/page.tsx
