+++ src/app/(unauthenticated)/layout.tsx
@@ -1,0 +1,1 @@
+export { ContentWrapper as default } from "@/components/content-wrapper";
